Refactors tests to reference parser via the builder
