+++ partiql-lang/src/test/kotlin/org/partiql/lang/eval/PartiqlAstExtensionsTests.kt
@@ -4,11 +4,12 @@ import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ArgumentsSource
 import org.partiql.lang.ast.SourceLocationMeta
 import org.partiql.lang.domains.PartiqlAst
-import org.partiql.lang.syntax.impl.PartiQLPigParser
+import org.partiql.lang.syntax.PartiQLParserBuilder
 import org.partiql.lang.util.ArgumentsProviderBase
 
 class PartiqlAstExtensionsTests : EvaluatorTestBase() {
-    private val parser = PartiQLPigParser()
+
+    private val parser = PartiQLParserBuilder.standard().build()
 
     data class StartingSourceLocationTestCase(val query: String, val expectedSourceLocationMeta: SourceLocationMeta)
