+++ partiql-lang/src/test/kotlin/org/partiql/lang/eval/evaluatortestframework/VisitorTransformBaseTestAdapter.kt
@@ -3,7 +3,7 @@ package org.partiql.lang.eval.evaluatortestframework
 import org.partiql.lang.CUSTOM_TEST_TYPES
 import org.partiql.lang.eval.EvaluationSession
 import org.partiql.lang.eval.visitors.VisitorTransformBase
-import org.partiql.lang.syntax.impl.PartiQLPigParser
+import org.partiql.lang.syntax.PartiQLParserBuilder
 import kotlin.test.assertEquals
 
 /** Test that the default visitor transformer returns an exact copy as the original ast */
@@ -19,9 +19,10 @@ class VisitorTransformBaseTestAdapter : EvaluatorTestAdapter {
     }
 
     private fun testVistorTransformBase(tc: EvaluatorTestDefinition) {
-        val parser = PartiQLPigParser(CUSTOM_TEST_TYPES)
+        val parser = PartiQLParserBuilder.standard()
+            .customTypes(CUSTOM_TEST_TYPES)
+            .build()
         val ast = parser.parseAstStatement(tc.query)
-
         val clonedAst = defaultTransformer.transformStatement(ast)
         assertEquals(
             ast,
