+++ partiql-lang/src/test/kotlin/org/partiql/lang/syntax/impl/PartiQLPigParserThreadInterruptTests.kt
@@ -1,4 +1,4 @@
-package org.partiql.lang.thread
+package org.partiql.lang.syntax.impl
 
 import com.amazon.ionelement.api.ionInt
 import io.mockk.every
@@ -17,17 +17,16 @@ import org.partiql.lang.StepContext
 import org.partiql.lang.domains.PartiqlAst
 import org.partiql.lang.eval.CompileOptions
 import org.partiql.lang.eval.visitors.VisitorTransformBase
-import org.partiql.lang.syntax.impl.PartiQLPigParser
 import org.partiql.parser.antlr.PartiQLTokens
 import java.io.InputStream
 import java.util.concurrent.atomic.AtomicBoolean
 import kotlin.concurrent.thread
 
-/** How long (in miilis) to wait after starting a thread to set the interrupted flag. */
-const val INTERRUPT_AFTER_MS: Long = 100
+/** How long (in millis) to wait after starting a thread to set the interrupted flag. */
+private const val INTERRUPT_AFTER_MS: Long = 100
 
 /** How long (in millis) to wait for a thread to terminate after setting the interrupted flag. */
-const val WAIT_FOR_THREAD_TERMINATION_MS: Long = 1000
+private const val WAIT_FOR_THREAD_TERMINATION_MS: Long = 1000
 
 /**
  * At various locations in this codebase we check the state of [Thread.interrupted] and throw an
@@ -39,7 +38,7 @@ const val WAIT_FOR_THREAD_TERMINATION_MS: Long = 1000
  */
 // Enforce execution of tests in same thread as we need the execution to be deterministic for interruption behavior.
 @Execution(ExecutionMode.SAME_THREAD)
-class ThreadInterruptedTests {
+class PartiQLPigParserThreadInterruptTests {
     private val bigPartiqlAst = makeBigPartiqlAstExpr(10000000)
 
     /**
